split random.jl into random/
