+++ base/random/normal.jl
@@ -1,868 +1,5 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-module Random
-
-using Base.dSFMT
-using Base.GMP: Limb, MPZ
-import Base: copymutable, copy, copy!, ==, hash
-
-export srand,
-       rand, rand!,
-       randn, randn!,
-       randexp, randexp!,
-       bitrand,
-       randstring,
-       randsubseq, randsubseq!,
-       shuffle, shuffle!,
-       randperm, randperm!,
-       randcycle, randcycle!,
-       AbstractRNG, MersenneTwister, RandomDevice,
-       GLOBAL_RNG, randjump
-
-
-abstract type AbstractRNG end
-
-abstract type FloatInterval end
-mutable struct CloseOpen <: FloatInterval end
-mutable struct Close1Open2 <: FloatInterval end
-
-
-## RandomDevice
-
-const BoolBitIntegerType = Union{Type{Bool},Base.BitIntegerType}
-const BoolBitIntegerArray = Union{Array{Bool},Base.BitIntegerArray}
-
-if Sys.iswindows()
-    struct RandomDevice <: AbstractRNG
-        buffer::Vector{UInt128}
-
-        RandomDevice() = new(Vector{UInt128}(1))
-    end
-
-    function rand(rd::RandomDevice, T::BoolBitIntegerType)
-        win32_SystemFunction036!(rd.buffer)
-        @inbounds return rd.buffer[1] % T
-    end
-
-    rand!(rd::RandomDevice, A::BoolBitIntegerArray) = (win32_SystemFunction036!(A); A)
-else # !windows
-    struct RandomDevice <: AbstractRNG
-        file::IOStream
-        unlimited::Bool
-
-        RandomDevice(unlimited::Bool=true) = new(open(unlimited ? "/dev/urandom" : "/dev/random"), unlimited)
-    end
-
-    rand(rd::RandomDevice, T::BoolBitIntegerType)   = read( rd.file, T)
-    rand!(rd::RandomDevice, A::BoolBitIntegerArray) = read!(rd.file, A)
-end # os-test
-
-"""
-    RandomDevice()
-
-Create a `RandomDevice` RNG object. Two such objects will always generate different streams of random numbers.
-"""
-RandomDevice
-
-
-rand(rng::RandomDevice, ::Type{Close1Open2}) =
-    reinterpret(Float64, 0x3ff0000000000000 | rand(rng, UInt64) & 0x000fffffffffffff)
-
-rand(rng::RandomDevice, ::Type{CloseOpen}) = rand(rng, Close1Open2) - 1.0
-
-
-## MersenneTwister
-
-const MTCacheLength = dsfmt_get_min_array_size()
-
-mutable struct MersenneTwister <: AbstractRNG
-    seed::Vector{UInt32}
-    state::DSFMT_state
-    vals::Vector{Float64}
-    idx::Int
-
-    function MersenneTwister(seed, state, vals, idx)
-        if !(length(vals) == MTCacheLength &&  0 <= idx <= MTCacheLength)
-            throw(DomainError(idx, "`length(vals)` and `idx` must be consistent with $MTCacheLength"))
-        end
-        new(seed, state, vals, idx)
-    end
-end
-
-MersenneTwister(seed::Vector{UInt32}, state::DSFMT_state) =
-    MersenneTwister(seed, state, zeros(Float64, MTCacheLength), MTCacheLength)
-
-"""
-    MersenneTwister(seed)
-
-Create a `MersenneTwister` RNG object. Different RNG objects can have their own seeds, which
-may be useful for generating different streams of random numbers.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> x1 = rand(rng, 2)
-2-element Array{Float64,1}:
- 0.590845
- 0.766797
-
-julia> rng = MersenneTwister(1234);
-
-julia> x2 = rand(rng, 2)
-2-element Array{Float64,1}:
- 0.590845
- 0.766797
-
-julia> x1 == x2
-true
-```
-"""
-MersenneTwister(seed) = srand(MersenneTwister(Vector{UInt32}(), DSFMT_state()), seed)
-
-function copy!(dst::MersenneTwister, src::MersenneTwister)
-    copy!(resize!(dst.seed, length(src.seed)), src.seed)
-    copy!(dst.state, src.state)
-    copy!(dst.vals, src.vals)
-    dst.idx = src.idx
-    dst
-end
-
-copy(src::MersenneTwister) =
-    MersenneTwister(copy(src.seed), copy(src.state), copy(src.vals), src.idx)
-
-==(r1::MersenneTwister, r2::MersenneTwister) =
-    r1.seed == r2.seed && r1.state == r2.state && isequal(r1.vals, r2.vals) && r1.idx == r2.idx
-
-hash(r::MersenneTwister, h::UInt) = foldr(hash, h, (r.seed, r.state, r.vals, r.idx))
-
-## Low level API for MersenneTwister
-
-@inline mt_avail(r::MersenneTwister) = MTCacheLength - r.idx
-@inline mt_empty(r::MersenneTwister) = r.idx == MTCacheLength
-@inline mt_setfull!(r::MersenneTwister) = r.idx = 0
-@inline mt_setempty!(r::MersenneTwister) = r.idx = MTCacheLength
-@inline mt_pop!(r::MersenneTwister) = @inbounds return r.vals[r.idx+=1]
-
-function gen_rand(r::MersenneTwister)
-    dsfmt_fill_array_close1_open2!(r.state, pointer(r.vals), length(r.vals))
-    mt_setfull!(r)
-end
-
-@inline reserve_1(r::MersenneTwister) = (mt_empty(r) && gen_rand(r); nothing)
-# `reserve` allows one to call `rand_inbounds` n times
-# precondition: n <= MTCacheLength
-@inline reserve(r::MersenneTwister, n::Int) = (mt_avail(r) < n && gen_rand(r); nothing)
-
-# precondition: !mt_empty(r)
-@inline rand_inbounds(r::MersenneTwister, ::Type{Close1Open2}) = mt_pop!(r)
-@inline rand_inbounds(r::MersenneTwister, ::Type{CloseOpen}) = rand_inbounds(r, Close1Open2) - 1.0
-@inline rand_inbounds(r::MersenneTwister) = rand_inbounds(r, CloseOpen)
-
-# produce Float64 values
-@inline rand(r::MersenneTwister, ::Type{I}) where {I<:FloatInterval} = (reserve_1(r); rand_inbounds(r, I))
-
-@inline rand_ui52_raw_inbounds(r::MersenneTwister) = reinterpret(UInt64, rand_inbounds(r, Close1Open2))
-@inline rand_ui52_raw(r::MersenneTwister) = (reserve_1(r); rand_ui52_raw_inbounds(r))
-
-@inline function rand_ui2x52_raw(r::MersenneTwister)
-    reserve(r, 2)
-    rand_ui52_raw_inbounds(r) % UInt128 << 64 | rand_ui52_raw_inbounds(r)
-end
-
-@inline function rand_ui104_raw(r::MersenneTwister)
-    reserve(r, 2)
-    rand_ui52_raw_inbounds(r) % UInt128 << 52 ⊻ rand_ui52_raw_inbounds(r)
-end
-
-function srand(r::MersenneTwister, seed::Vector{UInt32})
-    copy!(resize!(r.seed, length(seed)), seed)
-    dsfmt_init_by_array(r.state, r.seed)
-    mt_setempty!(r)
-    return r
-end
-
-# MersenneTwister jump
-
-"""
-    randjump(r::MersenneTwister, jumps::Integer, [jumppoly::AbstractString=dSFMT.JPOLY1e21]) -> Vector{MersenneTwister}
-
-Create an array of the size `jumps` of initialized `MersenneTwister` RNG objects. The
-first RNG object given as a parameter and following `MersenneTwister` RNGs in the array are
-initialized such that a state of the RNG object in the array would be moved forward (without
-generating numbers) from a previous RNG object array element on a particular number of steps
-encoded by the jump polynomial `jumppoly`.
-
-Default jump polynomial moves forward `MersenneTwister` RNG state by `10^20` steps.
-"""
-function randjump(mt::MersenneTwister, jumps::Integer, jumppoly::AbstractString)
-    mts = MersenneTwister[]
-    push!(mts, mt)
-    for i in 1:jumps-1
-        cmt = mts[end]
-        push!(mts, MersenneTwister(copy(cmt.seed), dSFMT.dsfmt_jump(cmt.state, jumppoly)))
-    end
-    return mts
-end
-randjump(r::MersenneTwister, jumps::Integer) = randjump(r, jumps, dSFMT.JPOLY1e21)
-
-## initialization
-
-function __init__()
-    try
-        srand()
-    catch ex
-        Base.showerror_nostdio(ex,
-            "WARNING: Error during initialization of module Random")
-    end
-end
-
-
-## make_seed()
-# make_seed methods produce values of type Array{UInt32}, suitable for MersenneTwister seeding
-
-function make_seed()
-    try
-        return rand(RandomDevice(), UInt32, 4)
-    catch
-        println(STDERR, "Entropy pool not available to seed RNG; using ad-hoc entropy sources.")
-        seed = reinterpret(UInt64, time())
-        seed = hash(seed, UInt64(getpid()))
-        try
-        seed = hash(seed, parse(UInt64, read(pipeline(`ifconfig`, `sha1sum`), String)[1:40], 16))
-        end
-        return make_seed(seed)
-    end
-end
-
-function make_seed(n::Integer)
-    n < 0 && throw(DomainError(n, "`n` must be non-negative."))
-    seed = UInt32[]
-    while true
-        push!(seed, n & 0xffffffff)
-        n >>= 32
-        if n == 0
-            return seed
-        end
-    end
-end
-
-## srand()
-
-"""
-    srand([rng=GLOBAL_RNG], seed) -> rng
-    srand([rng=GLOBAL_RNG]) -> rng
-
-Reseed the random number generator. If a `seed` is provided, the RNG will give a
-reproducible sequence of numbers, otherwise Julia will get entropy from the system. For
-`MersenneTwister`, the `seed` may be a non-negative integer or a vector of [`UInt32`](@ref)
-integers. `RandomDevice` does not support seeding.
-
-# Examples
-```jldoctest
-julia> srand(1234);
-
-julia> x1 = rand(2)
-2-element Array{Float64,1}:
- 0.590845
- 0.766797
-
-julia> srand(1234);
-
-julia> x2 = rand(2)
-2-element Array{Float64,1}:
- 0.590845
- 0.766797
-
-julia> x1 == x2
-true
-```
-"""
-srand(r::MersenneTwister) = srand(r, make_seed())
-srand(r::MersenneTwister, n::Integer) = srand(r, make_seed(n))
-
-function srand()
-    srand(GLOBAL_RNG)
-end
-
-function srand(seed::Union{Integer,Vector{UInt32}})
-    srand(GLOBAL_RNG, seed)
-end
-
-## Global RNG
-
-const GLOBAL_RNG = MersenneTwister(0)
-globalRNG() = GLOBAL_RNG
-
-# rand: a non-specified RNG defaults to GLOBAL_RNG
-
-"""
-    rand([rng=GLOBAL_RNG], [S], [dims...])
-
-Pick a random element or array of random elements from the set of values specified by `S`; `S` can be
-
-* an indexable collection (for example `1:n` or `['x','y','z']`),
-* an `Associative` or `AbstractSet` object,
-* a string (considered as a collection of characters), or
-* a type: the set of values to pick from is then equivalent to `typemin(S):typemax(S)` for
-  integers (this is not applicable to [`BigInt`](@ref)), and to ``[0, 1)`` for floating
-  point numbers;
-
-`S` defaults to [`Float64`](@ref).
-
-# Examples
-```julia-repl
-julia> rand(Int, 2)
-2-element Array{Int64,1}:
- 1339893410598768192
- 1575814717733606317
-
-julia> rand(MersenneTwister(0), Dict(1=>2, 3=>4))
-1=>2
-```
-
-!!! note
-    The complexity of `rand(rng, s::Union{Associative,AbstractSet})`
-    is linear in the length of `s`, unless an optimized method with
-    constant complexity is available, which is the case for `Dict`,
-    `Set` and `IntSet`. For more than a few calls, use `rand(rng,
-    collect(s))` instead, or either `rand(rng, Dict(s))` or `rand(rng,
-    Set(s))` as appropriate.
-"""
-@inline rand() = rand(GLOBAL_RNG, CloseOpen)
-@inline rand(T::Type) = rand(GLOBAL_RNG, T)
-rand(dims::Dims) = rand(GLOBAL_RNG, dims)
-rand(dims::Integer...) = rand(convert(Dims, dims))
-rand(T::Type, dims::Dims) = rand(GLOBAL_RNG, T, dims)
-rand(T::Type, d1::Integer, dims::Integer...) = rand(T, tuple(Int(d1), convert(Dims, dims)...))
-rand!(A::AbstractArray) = rand!(GLOBAL_RNG, A)
-
-rand(r::AbstractArray) = rand(GLOBAL_RNG, r)
-
-"""
-    rand!([rng=GLOBAL_RNG], A, [S=eltype(A)])
-
-Populate the array `A` with random values. If `S` is specified
-(`S` can be a type or a collection, cf. [`rand`](@ref) for details),
-the values are picked randomly from `S`.
-This is equivalent to `copy!(A, rand(rng, S, size(A)))`
-but without allocating a new array.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> rand!(rng, zeros(5))
-5-element Array{Float64,1}:
- 0.590845
- 0.766797
- 0.566237
- 0.460085
- 0.794026
-```
-"""
-rand!(A::AbstractArray, r::AbstractArray) = rand!(GLOBAL_RNG, A, r)
-
-rand(r::AbstractArray, dims::Dims) = rand(GLOBAL_RNG, r, dims)
-rand(r::AbstractArray, dims::Integer...) = rand(GLOBAL_RNG, r, convert(Dims, dims))
-
-## random floating point values
-
-@inline rand(r::AbstractRNG) = rand(r, CloseOpen)
-
-# MersenneTwister & RandomDevice
-@inline rand(r::Union{RandomDevice,MersenneTwister}, ::Type{Float64}) = rand(r, CloseOpen)
-
-rand_ui10_raw(r::MersenneTwister) = rand_ui52_raw(r)
-rand_ui23_raw(r::MersenneTwister) = rand_ui52_raw(r)
-rand_ui10_raw(r::AbstractRNG)    = rand(r, UInt16)
-rand_ui23_raw(r::AbstractRNG)    = rand(r, UInt32)
-
-rand(r::Union{RandomDevice,MersenneTwister}, ::Type{Float16}) =
-    Float16(reinterpret(Float32, (rand_ui10_raw(r) % UInt32 << 13) & 0x007fe000 | 0x3f800000) - 1)
-
-rand(r::Union{RandomDevice,MersenneTwister}, ::Type{Float32}) =
-    reinterpret(Float32, rand_ui23_raw(r) % UInt32 & 0x007fffff | 0x3f800000) - 1
-
-
-## random integers
-
-@inline rand_ui52_raw(r::AbstractRNG) = reinterpret(UInt64, rand(r, Close1Open2))
-@inline rand_ui52(r::AbstractRNG) = rand_ui52_raw(r) & 0x000fffffffffffff
-
-# MersenneTwister
-
-@inline rand(r::MersenneTwister, ::Type{T}) where {T<:Union{Bool,Int8,UInt8,Int16,UInt16,Int32,UInt32}} =
-    rand_ui52_raw(r) % T
-
-function rand(r::MersenneTwister, ::Type{UInt64})
-    reserve(r, 2)
-    rand_ui52_raw_inbounds(r) << 32 ⊻ rand_ui52_raw_inbounds(r)
-end
-
-function rand(r::MersenneTwister, ::Type{UInt128})
-    reserve(r, 3)
-    xor(rand_ui52_raw_inbounds(r) % UInt128 << 96,
-        rand_ui52_raw_inbounds(r) % UInt128 << 48,
-        rand_ui52_raw_inbounds(r))
-end
-
-rand(r::MersenneTwister, ::Type{Int64})   = reinterpret(Int64,  rand(r, UInt64))
-rand(r::MersenneTwister, ::Type{Int128})  = reinterpret(Int128, rand(r, UInt128))
-
-## random Complex values
-
-rand(r::AbstractRNG, ::Type{Complex{T}}) where {T<:Real} = complex(rand(r, T), rand(r, T))
-
-# random Char values
-# returns a random valid Unicode scalar value (i.e. 0 - 0xd7ff, 0xe000 - # 0x10ffff)
-function rand(r::AbstractRNG, ::Type{Char})
-    c = rand(r, 0x00000000:0x0010f7ff)
-    (c < 0xd800) ? Char(c) : Char(c+0x800)
-end
-
-# random values from Dict, Set, IntSet (for efficiency)
-function rand(r::AbstractRNG, t::Dict)
-    isempty(t) && throw(ArgumentError("collection must be non-empty"))
-    rg = RangeGenerator(1:length(t.slots))
-    while true
-        i = rand(r, rg)
-        Base.isslotfilled(t, i) && @inbounds return (t.keys[i] => t.vals[i])
-    end
-end
-
-rand(r::AbstractRNG, s::Set) = rand(r, s.dict).first
-
-function rand(r::AbstractRNG, s::IntSet)
-    isempty(s) && throw(ArgumentError("collection must be non-empty"))
-    # s can be empty while s.bits is not, so we cannot rely on the
-    # length check in RangeGenerator below
-    rg = RangeGenerator(1:length(s.bits))
-    while true
-        n = rand(r, rg)
-        @inbounds b = s.bits[n]
-        b && return n
-    end
-end
-
-function nth(iter, n::Integer)::eltype(iter)
-    for (i, x) in enumerate(iter)
-        i == n && return x
-    end
-end
-nth(iter::AbstractArray, n::Integer) = iter[n]
-
-rand(r::AbstractRNG, s::Union{Associative,AbstractSet}) = nth(s, rand(r, 1:length(s)))
-
-rand(s::Union{Associative,AbstractSet}) = rand(GLOBAL_RNG, s)
-
-## Arrays of random numbers
-
-rand(r::AbstractRNG, dims::Dims) = rand(r, Float64, dims)
-rand(r::AbstractRNG, dims::Integer...) = rand(r, convert(Dims, dims))
-
-rand(r::AbstractRNG, T::Type, dims::Dims) = rand!(r, Array{T}(dims))
-rand(r::AbstractRNG, T::Type, d1::Integer, dims::Integer...) = rand(r, T, tuple(Int(d1), convert(Dims, dims)...))
-# note: the above method would trigger an ambiguity warning if d1 was not separated out:
-# rand(r, ()) would match both this method and rand(r, dims::Dims)
-# moreover, a call like rand(r, NotImplementedType()) would be an infinite loop
-
-function rand!(r::AbstractRNG, A::AbstractArray{T}, ::Type{X}=T) where {T,X}
-    for i in eachindex(A)
-        @inbounds A[i] = rand(r, X)
-    end
-    A
-end
-
-rand!(A::AbstractArray, ::Type{X}) where {X} = rand!(GLOBAL_RNG, A, X)
-
-function rand!(r::AbstractRNG, A::AbstractArray, s::Union{Dict,Set,IntSet})
-    for i in eachindex(A)
-        @inbounds A[i] = rand(r, s)
-    end
-    A
-end
-
-# avoid linear complexity for repeated calls with generic containers
-rand!(r::AbstractRNG, A::AbstractArray, s::Union{Associative,AbstractSet}) = rand!(r, A, collect(s))
-
-rand!(A::AbstractArray, s::Union{Associative,AbstractSet}) = rand!(GLOBAL_RNG, A, s)
-
-rand(r::AbstractRNG, s::Associative{K,V}, dims::Dims) where {K,V} = rand!(r, Array{Pair{K,V}}(dims), s)
-rand(r::AbstractRNG, s::AbstractSet{T}, dims::Dims) where {T} = rand!(r, Array{T}(dims), s)
-rand(r::AbstractRNG, s::Union{Associative,AbstractSet}, dims::Integer...) = rand(r, s, convert(Dims, dims))
-rand(s::Union{Associative,AbstractSet}, dims::Integer...) = rand(GLOBAL_RNG, s, convert(Dims, dims))
-rand(s::Union{Associative,AbstractSet}, dims::Dims) = rand(GLOBAL_RNG, s, dims)
-
-# MersenneTwister
-
-function rand_AbstractArray_Float64!(r::MersenneTwister, A::AbstractArray{Float64},
-                                     n=length(A), ::Type{I}=CloseOpen) where I<:FloatInterval
-    # what follows is equivalent to this simple loop but more efficient:
-    # for i=1:n
-    #     @inbounds A[i] = rand(r, I)
-    # end
-    m = 0
-    while m < n
-        s = mt_avail(r)
-        if s == 0
-            gen_rand(r)
-            s = mt_avail(r)
-        end
-        m2 = min(n, m+s)
-        for i=m+1:m2
-            @inbounds A[i] = rand_inbounds(r, I)
-        end
-        m = m2
-    end
-    A
-end
-
-rand!(r::MersenneTwister, A::AbstractArray{Float64}) = rand_AbstractArray_Float64!(r, A)
-
-fill_array!(s::DSFMT_state, A::Ptr{Float64}, n::Int, ::Type{CloseOpen}) = dsfmt_fill_array_close_open!(s, A, n)
-fill_array!(s::DSFMT_state, A::Ptr{Float64}, n::Int, ::Type{Close1Open2}) = dsfmt_fill_array_close1_open2!(s, A, n)
-
-function rand!(r::MersenneTwister, A::Array{Float64}, n::Int=length(A), ::Type{I}=CloseOpen) where I<:FloatInterval
-    # depending on the alignment of A, the data written by fill_array! may have
-    # to be left-shifted by up to 15 bytes (cf. unsafe_copy! below) for
-    # reproducibility purposes;
-    # so, even for well aligned arrays, fill_array! is used to generate only
-    # the n-2 first values (or n-3 if n is odd), and the remaining values are
-    # generated by the scalar version of rand
-    if n > length(A)
-        throw(BoundsError(A,n))
-    end
-    n2 = (n-2) ÷ 2 * 2
-    if n2 < dsfmt_get_min_array_size()
-        rand_AbstractArray_Float64!(r, A, n, I)
-    else
-        pA = pointer(A)
-        align = Csize_t(pA) % 16
-        if align > 0
-            pA2 = pA + 16 - align
-            fill_array!(r.state, pA2, n2, I) # generate the data in-place, but shifted
-            unsafe_copy!(pA, pA2, n2) # move the data to the beginning of the array
-        else
-            fill_array!(r.state, pA, n2, I)
-        end
-        for i=n2+1:n
-            @inbounds A[i] = rand(r, I)
-        end
-    end
-    A
-end
-
-@inline mask128(u::UInt128, ::Type{Float16}) = (u & 0x03ff03ff03ff03ff03ff03ff03ff03ff) | 0x3c003c003c003c003c003c003c003c00
-@inline mask128(u::UInt128, ::Type{Float32}) = (u & 0x007fffff007fffff007fffff007fffff) | 0x3f8000003f8000003f8000003f800000
-
-function rand!(r::MersenneTwister, A::Union{Array{Float16},Array{Float32}}, ::Type{Close1Open2})
-    T = eltype(A)
-    n = length(A)
-    n128 = n * sizeof(T) ÷ 16
-    rand!(r, unsafe_wrap(Array, convert(Ptr{Float64}, pointer(A)), 2*n128), 2*n128, Close1Open2)
-    A128 = unsafe_wrap(Array, convert(Ptr{UInt128}, pointer(A)), n128)
-    @inbounds for i in 1:n128
-        u = A128[i]
-        u ⊻= u << 26
-        # at this point, the 64 low bits of u, "k" being the k-th bit of A128[i] and "+" the bit xor, are:
-        # [..., 58+32,..., 53+27, 52+26, ..., 33+7, 32+6, ..., 27+1, 26, ..., 1]
-        # the bits needing to be random are
-        # [1:10, 17:26, 33:42, 49:58] (for Float16)
-        # [1:23, 33:55] (for Float32)
-        # this is obviously satisfied on the 32 low bits side, and on the high side, the entropy comes
-        # from bits 33:52 of A128[i] and then from bits 27:32 (which are discarded on the low side)
-        # this is similar for the 64 high bits of u
-        A128[i] = mask128(u, T)
-    end
-    for i in 16*n128÷sizeof(T)+1:n
-        @inbounds A[i] = rand(r, T) + oneunit(T)
-    end
-    A
-end
-
-function rand!(r::MersenneTwister, A::Union{Array{Float16},Array{Float32}}, ::Type{CloseOpen})
-    rand!(r, A, Close1Open2)
-    I32 = one(Float32)
-    for i in eachindex(A)
-        @inbounds A[i] = Float32(A[i])-I32 # faster than "A[i] -= one(T)" for T==Float16
-    end
-    A
-end
-
-rand!(r::MersenneTwister, A::Union{Array{Float16},Array{Float32}}) = rand!(r, A, CloseOpen)
-
-function rand!(r::MersenneTwister, A::Array{UInt128}, n::Int=length(A))
-    if n > length(A)
-        throw(BoundsError(A,n))
-    end
-    Af = unsafe_wrap(Array, convert(Ptr{Float64}, pointer(A)), 2n)
-    i = n
-    while true
-        rand!(r, Af, 2i, Close1Open2)
-        n < 5 && break
-        i = 0
-        @inbounds while n-i >= 5
-            u = A[i+=1]
-            A[n]    ⊻= u << 48
-            A[n-=1] ⊻= u << 36
-            A[n-=1] ⊻= u << 24
-            A[n-=1] ⊻= u << 12
-            n-=1
-        end
-    end
-    if n > 0
-        u = rand_ui2x52_raw(r)
-        for i = 1:n
-            @inbounds A[i] ⊻= u << (12*i)
-        end
-    end
-    A
-end
-
-# A::Array{UInt128} will match the specialized method above
-function rand!(r::MersenneTwister, A::Base.BitIntegerArray)
-    n = length(A)
-    T = eltype(A)
-    n128 = n * sizeof(T) ÷ 16
-    rand!(r, unsafe_wrap(Array, convert(Ptr{UInt128}, pointer(A)), n128))
-    for i = 16*n128÷sizeof(T)+1:n
-        @inbounds A[i] = rand(r, T)
-    end
-    A
-end
-
-## Generate random integer within a range
-
-# remainder function according to Knuth, where rem_knuth(a, 0) = a
-rem_knuth(a::UInt, b::UInt) = a % (b + (b == 0)) + a * (b == 0)
-rem_knuth(a::T, b::T) where {T<:Unsigned} = b != 0 ? a % b : a
-
-# maximum multiple of k <= 2^bits(T) decremented by one,
-# that is 0xFFFF...FFFF if k = typemax(T) - typemin(T) with intentional underflow
-# see http://stackoverflow.com/questions/29182036/integer-arithmetic-add-1-to-uint-max-and-divide-by-n-without-overflow
-maxmultiple(k::T) where {T<:Unsigned} = (div(typemax(T) - k + oneunit(k), k + (k == 0))*k + k - oneunit(k))::T
-
-# maximum multiple of k within 1:2^32 or 1:2^64 decremented by one, depending on size
-maxmultiplemix(k::UInt64) = if k >> 32 != 0; maxmultiple(k); else (div(0x0000000100000000, k + (k == 0))*k - oneunit(k))::UInt64; end
-
-abstract type RangeGenerator end
-
-struct RangeGeneratorInt{T<:Integer,U<:Unsigned} <: RangeGenerator
-    a::T   # first element of the range
-    k::U   # range length or zero for full range
-    u::U   # rejection threshold
-end
-# generators with 32, 128 bits entropy
-RangeGeneratorInt(a::T, k::U) where {T,U<:Union{UInt32,UInt128}} = RangeGeneratorInt{T,U}(a, k, maxmultiple(k))
-# mixed 32/64 bits entropy generator
-RangeGeneratorInt(a::T, k::UInt64) where {T} = RangeGeneratorInt{T,UInt64}(a, k, maxmultiplemix(k))
-# generator for ranges
-function RangeGenerator(r::UnitRange{T}) where T<:Unsigned
-    if isempty(r)
-        throw(ArgumentError("range must be non-empty"))
-    end
-    RangeGeneratorInt(first(r), last(r) - first(r) + oneunit(T))
-end
-
-# specialized versions
-for (T, U) in [(UInt8, UInt32), (UInt16, UInt32),
-               (Int8, UInt32), (Int16, UInt32), (Int32, UInt32), (Int64, UInt64), (Int128, UInt128),
-               (Bool, UInt32)]
-
-    @eval RangeGenerator(r::UnitRange{$T}) = begin
-        if isempty(r)
-            throw(ArgumentError("range must be non-empty"))
-        end
-        RangeGeneratorInt(first(r), convert($U, unsigned(last(r) - first(r)) + one($U))) # overflow ok
-    end
-end
-
-struct RangeGeneratorBigInt <: RangeGenerator
-    a::BigInt         # first
-    m::BigInt         # range length - 1
-    nlimbs::Int       # number of limbs in generated BigInt's (z ∈ [0, m])
-    nlimbsmax::Int    # max number of limbs for z+a
-    mask::Limb        # applied to the highest limb
-end
-
-
-function RangeGenerator(r::UnitRange{BigInt})
-    m = last(r) - first(r)
-    m < 0 && throw(ArgumentError("range must be non-empty"))
-    nd = ndigits(m, 2)
-    nlimbs, highbits = divrem(nd, 8*sizeof(Limb))
-    highbits > 0 && (nlimbs += 1)
-    mask = highbits == 0 ? ~zero(Limb) : one(Limb)<<highbits - one(Limb)
-    nlimbsmax = max(nlimbs, abs(last(r).size), abs(first(r).size))
-    return RangeGeneratorBigInt(first(r), m, nlimbs, nlimbsmax, mask)
-end
-
-
-# this function uses 32 bit entropy for small ranges of length <= typemax(UInt32) + 1
-# RangeGeneratorInt is responsible for providing the right value of k
-function rand(rng::AbstractRNG, g::RangeGeneratorInt{T,UInt64}) where T<:Union{UInt64,Int64}
-    local x::UInt64
-    if (g.k - 1) >> 32 == 0
-        x = rand(rng, UInt32)
-        while x > g.u
-            x = rand(rng, UInt32)
-        end
-    else
-        x = rand(rng, UInt64)
-        while x > g.u
-            x = rand(rng, UInt64)
-        end
-    end
-    return reinterpret(T, reinterpret(UInt64, g.a) + rem_knuth(x, g.k))
-end
-
-function rand(rng::AbstractRNG, g::RangeGeneratorInt{T,U}) where U<:Unsigned where T<:Integer
-    x = rand(rng, U)
-    while x > g.u
-        x = rand(rng, U)
-    end
-    (unsigned(g.a) + rem_knuth(x, g.k)) % T
-end
-
-function rand(rng::AbstractRNG, g::RangeGeneratorBigInt)
-    x = MPZ.realloc2(g.nlimbsmax*8*sizeof(Limb))
-    limbs = unsafe_wrap(Array, x.d, g.nlimbs)
-    while true
-        rand!(rng, limbs)
-        @inbounds limbs[end] &= g.mask
-        MPZ.mpn_cmp(x, g.m, g.nlimbs) <= 0 && break
-    end
-    # adjust x.size (normally done by mpz_limbs_finish, in GMP version >= 6)
-    x.size = g.nlimbs
-    while x.size > 0
-        @inbounds limbs[x.size] != 0 && break
-        x.size -= 1
-    end
-    MPZ.add!(x, g.a)
-end
-
-rand(rng::AbstractRNG, r::UnitRange{<:Union{Signed,Unsigned,BigInt,Bool}}) = rand(rng, RangeGenerator(r))
-
-## special case for MersenneTwister
-@inline function rand_lteq(r::AbstractRNG, randfun, u::U, mask::U) where U<:Integer
-    while true
-        x = randfun(r) & mask
-        x <= u && return x
-    end
-end
-
-function rand(rng::MersenneTwister, r::UnitRange{T}) where T<:Union{Base.BitInteger64,Bool}
-    isempty(r) && throw(ArgumentError("range must be non-empty"))
-    m = last(r) % UInt64 - first(r) % UInt64
-    bw = (64 - leading_zeros(m)) % UInt # bit-width
-    mask = (1 % UInt64 << bw) - (1 % UInt64)
-    x = bw <= 52 ? rand_lteq(rng, rand_ui52_raw, m, mask) :
-                   rand_lteq(rng, rng->rand(rng, UInt64), m, mask)
-    (x + first(r) % UInt64) % T
-end
-
-function rand(rng::MersenneTwister, r::UnitRange{T}) where T<:Union{Int128,UInt128}
-    isempty(r) && throw(ArgumentError("range must be non-empty"))
-    m = (last(r)-first(r)) % UInt128
-    bw = (128 - leading_zeros(m)) % UInt # bit-width
-    mask = (1 % UInt128 << bw) - (1 % UInt128)
-    x = bw <= 52  ? rand_lteq(rng, rand_ui52_raw, m % UInt64, mask % UInt64) % UInt128 :
-        bw <= 104 ? rand_lteq(rng, rand_ui104_raw, m, mask) :
-                    rand_lteq(rng, rng->rand(rng, UInt128), m, mask)
-    x % T + first(r)
-end
-
-# Randomly draw a sample from an AbstractArray r
-# (e.g. r is a range 0:2:8 or a vector [2, 3, 5, 7])
-rand(rng::AbstractRNG, r::AbstractArray) = @inbounds return r[rand(rng, 1:length(r))]
-
-function rand!(rng::AbstractRNG, A::AbstractArray, g::RangeGenerator)
-    for i in eachindex(A)
-        @inbounds A[i] = rand(rng, g)
-    end
-    return A
-end
-
-rand!(rng::AbstractRNG, A::AbstractArray, r::UnitRange{<:Union{Signed,Unsigned,BigInt,Bool,Char}}) = rand!(rng, A, RangeGenerator(r))
-
-function rand!(rng::AbstractRNG, A::AbstractArray, r::AbstractArray)
-    g = RangeGenerator(1:(length(r)))
-    for i in eachindex(A)
-        @inbounds A[i] = r[rand(rng, g)]
-    end
-    return A
-end
-
-rand(rng::AbstractRNG, r::AbstractArray{T}, dims::Dims) where {T} = rand!(rng, Array{T}(dims), r)
-rand(rng::AbstractRNG, r::AbstractArray, dims::Integer...) = rand(rng, r, convert(Dims, dims))
-
-# rand from a string
-
-isvalid_unsafe(s::String, i) = !Base.is_valid_continuation(unsafe_load(pointer(s), i))
-isvalid_unsafe(s::AbstractString, i) = isvalid(s, i)
-_endof(s::String) = sizeof(s)
-_endof(s::AbstractString) = endof(s)
-
-function rand(rng::AbstractRNG, s::AbstractString)::Char
-    g = RangeGenerator(1:_endof(s))
-    while true
-        pos = rand(rng, g)
-        isvalid_unsafe(s, pos) && return s[pos]
-    end
-end
-
-rand(s::AbstractString) = rand(GLOBAL_RNG, s)
-
-## rand from a string for arrays
-# we use collect(str), which is most of the time more efficient than specialized methods
-# (except maybe for very small arrays)
-rand!(rng::AbstractRNG, A::AbstractArray, str::AbstractString) = rand!(rng, A, collect(str))
-rand!(A::AbstractArray, str::AbstractString) = rand!(GLOBAL_RNG, A, str)
-rand(rng::AbstractRNG, str::AbstractString, dims::Dims) = rand!(rng, Array{eltype(str)}(dims), str)
-rand(rng::AbstractRNG, str::AbstractString, d1::Integer, dims::Integer...) = rand(rng, str, convert(Dims, tuple(d1, dims...)))
-rand(str::AbstractString, dims::Dims) = rand(GLOBAL_RNG, str, dims)
-rand(str::AbstractString, d1::Integer, dims::Integer...) = rand(GLOBAL_RNG, str, d1, dims...)
-
-## random BitArrays (AbstractRNG)
-
-function rand!(rng::AbstractRNG, B::BitArray)
-    isempty(B) && return B
-    Bc = B.chunks
-    rand!(rng, Bc)
-    Bc[end] &= Base._msk_end(B)
-    return B
-end
-
-"""
-    bitrand([rng=GLOBAL_RNG], [dims...])
-
-Generate a `BitArray` of random boolean values.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> bitrand(rng, 10)
-10-element BitArray{1}:
-  true
-  true
-  true
- false
-  true
- false
- false
-  true
- false
-  true
-```
-"""
-bitrand(r::AbstractRNG, dims::Dims)   = rand!(r, BitArray(dims))
-bitrand(r::AbstractRNG, dims::Integer...) = rand!(r, BitArray(convert(Dims, dims)))
-
-bitrand(dims::Dims)   = rand!(BitArray(dims))
-bitrand(dims::Integer...) = rand!(BitArray(convert(Dims, dims)))
-
 ## randn() - Normally distributed random numbers using Ziggurat algorithm
 
 # The Ziggurat Method for generating random variables - Marsaglia and Tsang
@@ -1524,417 +661,3 @@ end
 Base.@irrational SQRT_HALF 0.7071067811865475244008  sqrt(big(0.5))
 randn(rng::AbstractRNG, ::Type{Complex{T}}) where {T<:AbstractFloat} =
     Complex{T}(SQRT_HALF * randn(rng, T), SQRT_HALF * randn(rng, T))
-
-## random UUID generation
-
-struct UUID
-    value::UInt128
-
-    UUID(u::UInt128) = new(u)
-end
-
-"""
-    uuid1([rng::AbstractRNG=GLOBAL_RNG]) -> UUID
-
-Generates a version 1 (time-based) universally unique identifier (UUID), as specified
-by RFC 4122. Note that the Node ID is randomly generated (does not identify the host)
-according to section 4.5 of the RFC.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> Base.Random.uuid1(rng)
-2cc938da-5937-11e7-196e-0f4ef71aa64b
-```
-"""
-function uuid1(rng::AbstractRNG=GLOBAL_RNG)
-    u = rand(rng, UInt128)
-
-    # mask off clock sequence and node
-    u &= 0x00000000000000003fffffffffffffff
-
-    # set the unicast/multicast bit and version
-    u |= 0x00000000000010000000010000000000
-
-    # 0x01b21dd213814000 is the number of 100 nanosecond intervals
-    # between the UUID epoch and Unix epoch
-    timestamp = round(UInt64, time() * 1e7) + 0x01b21dd213814000
-    ts_low = timestamp & typemax(UInt32)
-    ts_mid = (timestamp >> 32) & typemax(UInt16)
-    ts_hi = (timestamp >> 48) & 0x0fff
-
-    u |= UInt128(ts_low) << 96
-    u |= UInt128(ts_mid) << 80
-    u |= UInt128(ts_hi) << 64
-
-    UUID(u)
-end
-
-"""
-    uuid4([rng::AbstractRNG=GLOBAL_RNG]) -> UUID
-
-Generates a version 4 (random or pseudo-random) universally unique identifier (UUID),
-as specified by RFC 4122.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> Base.Random.uuid4(rng)
-82015f10-44cc-4827-996e-0f4ef71aa64b
-```
-"""
-function uuid4(rng::AbstractRNG=GLOBAL_RNG)
-    u = rand(rng, UInt128)
-    u &= 0xffffffffffff0fff3fffffffffffffff
-    u |= 0x00000000000040008000000000000000
-    UUID(u)
-end
-
-"""
-    uuid_version(u::UUID) -> Integer
-
-Inspects the given UUID and returns its version (see RFC 4122).
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> Base.Random.uuid_version(Base.Random.uuid4(rng))
-4
-```
-"""
-function uuid_version(u::UUID)
-    Int((u.value >> 76) & 0xf)
-end
-
-Base.convert(::Type{UInt128}, u::UUID) = u.value
-
-function Base.convert(::Type{UUID}, s::AbstractString)
-    s = lowercase(s)
-
-    if !ismatch(r"^[0-9a-f]{8}(?:-[0-9a-f]{4}){3}-[0-9a-f]{12}$", s)
-        throw(ArgumentError("Malformed UUID string"))
-    end
-
-    u = UInt128(0)
-    for i in [1:8; 10:13; 15:18; 20:23; 25:36]
-        u <<= 4
-        d = s[i]-'0'
-        u |= 0xf & (d-39*(d>9))
-    end
-    return UUID(u)
-end
-
-function Base.repr(u::UUID)
-    u = u.value
-    a = Vector{UInt8}(36)
-    for i = [36:-1:25; 23:-1:20; 18:-1:15; 13:-1:10; 8:-1:1]
-        d = u & 0xf
-        a[i] = '0'+d+39*(d>9)
-        u >>= 4
-    end
-    a[[24,19,14,9]] = '-'
-
-    return String(a)
-end
-
-Base.show(io::IO, u::UUID) = write(io, Base.repr(u))
-
-# return a random string (often useful for temporary filenames/dirnames)
-
-"""
-    randstring([rng=GLOBAL_RNG], [chars], [len=8])
-
-Create a random string of length `len`, consisting of characters from
-`chars`, which defaults to the set of upper- and lower-case letters
-and the digits 0-9. The optional `rng` argument specifies a random
-number generator, see [Random Numbers](@ref).
-
-# Examples
-```jldoctest
-julia> srand(0); randstring()
-"c03rgKi1"
-
-julia> randstring(MersenneTwister(0), 'a':'z', 6)
-"wijzek"
-
-julia> randstring("ACGT")
-"TATCGGTC"
-```
-
-!!! note
-    `chars` can be any collection of characters, of type `Char` or
-    `UInt8` (more efficient), provided [`rand`](@ref) can randomly
-    pick characters from it.
-"""
-function randstring end
-
-let b = UInt8['0':'9';'A':'Z';'a':'z']
-    global randstring
-    randstring(r::AbstractRNG, chars=b, n::Integer=8) = String(rand(r, chars, n))
-    randstring(r::AbstractRNG, n::Integer) = randstring(r, b, n)
-    randstring(chars=b, n::Integer=8) = randstring(GLOBAL_RNG, chars, n)
-    randstring(n::Integer) = randstring(GLOBAL_RNG, b, n)
-end
-
-# Fill S (resized as needed) with a random subsequence of A, where
-# each element of A is included in S with independent probability p.
-# (Note that this is different from the problem of finding a random
-#  size-m subset of A where m is fixed!)
-
-"""
-    randsubseq!(S, A, p)
-
-Like [`randsubseq`](@ref), but the results are stored in `S`
-(which is resized as needed).
-"""
-function randsubseq!(r::AbstractRNG, S::AbstractArray, A::AbstractArray, p::Real)
-    0 <= p <= 1 || throw(ArgumentError("probability $p not in [0,1]"))
-    n = length(A)
-    p == 1 && return copy!(resize!(S, n), A)
-    empty!(S)
-    p == 0 && return S
-    nexpected = p * length(A)
-    sizehint!(S, round(Int,nexpected + 5*sqrt(nexpected)))
-    if p > 0.15 # empirical threshold for trivial O(n) algorithm to be better
-        for i = 1:n
-            rand(r) <= p && push!(S, A[i])
-        end
-    else
-        # Skip through A, in order, from each element i to the next element i+s
-        # included in S. The probability that the next included element is
-        # s==k (k > 0) is (1-p)^(k-1) * p, and hence the probability (CDF) that
-        # s is in {1,...,k} is 1-(1-p)^k = F(k).   Thus, we can draw the skip s
-        # from this probability distribution via the discrete inverse-transform
-        # method: s = ceil(F^{-1}(u)) where u = rand(), which is simply
-        # s = ceil(log(rand()) / log1p(-p)).
-        # -log(rand()) is an exponential variate, so can use randexp().
-        L = -1 / log1p(-p) # L > 0
-        i = 0
-        while true
-            s = randexp(r) * L
-            s >= n - i && return S # compare before ceil to avoid overflow
-            push!(S, A[i += ceil(Int,s)])
-        end
-        # [This algorithm is similar in spirit to, but much simpler than,
-        #  the one by Vitter for a related problem in "Faster methods for
-        #  random sampling," Comm. ACM Magazine 7, 703-718 (1984).]
-    end
-    return S
-end
-randsubseq!(S::AbstractArray, A::AbstractArray, p::Real) = randsubseq!(GLOBAL_RNG, S, A, p)
-
-randsubseq(r::AbstractRNG, A::AbstractArray{T}, p::Real) where {T} = randsubseq!(r, T[], A, p)
-
-"""
-    randsubseq(A, p) -> Vector
-
-Return a vector consisting of a random subsequence of the given array `A`, where each
-element of `A` is included (in order) with independent probability `p`. (Complexity is
-linear in `p*length(A)`, so this function is efficient even if `p` is small and `A` is
-large.) Technically, this process is known as "Bernoulli sampling" of `A`.
-"""
-randsubseq(A::AbstractArray, p::Real) = randsubseq(GLOBAL_RNG, A, p)
-
-"Return a random `Int` (masked with `mask`) in ``[0, n)``, when `n <= 2^52`."
-@inline function rand_lt(r::AbstractRNG, n::Int, mask::Int=nextpow2(n)-1)
-    # this duplicates the functionality of RangeGenerator objects,
-    # to optimize this special case
-    while true
-        x = (rand_ui52_raw(r) % Int) & mask
-        x < n && return x
-    end
-end
-
-"""
-    shuffle!([rng=GLOBAL_RNG,] v::AbstractArray)
-
-In-place version of [`shuffle`](@ref): randomly permute `v` in-place,
-optionally supplying the random-number generator `rng`.
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> shuffle!(rng, collect(1:16))
-16-element Array{Int64,1}:
-  2
- 15
-  5
- 14
-  1
-  9
- 10
-  6
- 11
-  3
- 16
-  7
-  4
- 12
-  8
- 13
-```
-"""
-function shuffle!(r::AbstractRNG, a::AbstractArray)
-    n = length(a)
-    @assert n <= Int64(2)^52
-    mask = nextpow2(n) - 1
-    for i = n:-1:2
-        (mask >> 1) == i && (mask >>= 1)
-        j = 1 + rand_lt(r, i, mask)
-        a[i], a[j] = a[j], a[i]
-    end
-    return a
-end
-
-shuffle!(a::AbstractArray) = shuffle!(GLOBAL_RNG, a)
-
-"""
-    shuffle([rng=GLOBAL_RNG,] v::AbstractArray)
-
-Return a randomly permuted copy of `v`. The optional `rng` argument specifies a random
-number generator (see [Random Numbers](@ref)).
-To permute `v` in-place, see [`shuffle!`](@ref). To obtain randomly permuted
-indices, see [`randperm`](@ref).
-
-# Examples
-```jldoctest
-julia> rng = MersenneTwister(1234);
-
-julia> shuffle(rng, collect(1:10))
-10-element Array{Int64,1}:
-  6
-  1
- 10
-  2
-  3
-  9
-  5
-  7
-  4
-  8
-```
-"""
-shuffle(r::AbstractRNG, a::AbstractArray) = shuffle!(r, copymutable(a))
-shuffle(a::AbstractArray) = shuffle(GLOBAL_RNG, a)
-
-"""
-    randperm([rng=GLOBAL_RNG,] n::Integer)
-
-Construct a random permutation of length `n`. The optional `rng` argument specifies a random
-number generator (see [Random Numbers](@ref)).
-To randomly permute an arbitrary vector, see [`shuffle`](@ref)
-or [`shuffle!`](@ref).
-
-# Examples
-```jldoctest
-julia> randperm(MersenneTwister(1234), 4)
-4-element Array{Int64,1}:
- 2
- 1
- 4
- 3
-```
-"""
-randperm(r::AbstractRNG, n::Integer) = randperm!(r, Vector{Int}(n))
-randperm(n::Integer) = randperm(GLOBAL_RNG, n)
-
-"""
-    randperm!([rng=GLOBAL_RNG,] A::Array{<:Integer})
-
-Construct in `A` a random permutation of length `length(A)`. The
-optional `rng` argument specifies a random number generator (see
-[Random Numbers](@ref)). To randomly permute an arbitrary vector, see
-[`shuffle`](@ref) or [`shuffle!`](@ref).
-
-# Examples
-```jldoctest
-julia> randperm!(MersenneTwister(1234), Vector{Int}(4))
-4-element Array{Int64,1}:
- 2
- 1
- 4
- 3
-```
-"""
-function randperm!(r::AbstractRNG, a::Array{<:Integer})
-    n = length(a)
-    @assert n <= Int64(2)^52
-    n == 0 && return a
-    a[1] = 1
-    mask = 3
-    @inbounds for i = 2:n
-        j = 1 + rand_lt(r, i, mask)
-        if i != j # a[i] is uninitialized (and could be #undef)
-            a[i] = a[j]
-        end
-        a[j] = i
-        i == 1+mask && (mask = 2mask + 1)
-    end
-    return a
-end
-
-randperm!(a::Array{<:Integer}) = randperm!(GLOBAL_RNG, a)
-
-
-"""
-    randcycle([rng=GLOBAL_RNG,] n::Integer)
-
-Construct a random cyclic permutation of length `n`. The optional `rng`
-argument specifies a random number generator, see [Random Numbers](@ref).
-
-# Examples
-```jldoctest
-julia> randcycle(MersenneTwister(1234), 6)
-6-element Array{Int64,1}:
- 3
- 5
- 4
- 6
- 1
- 2
-```
-"""
-randcycle(r::AbstractRNG, n::Integer) = randcycle!(r, Vector{Int}(n))
-randcycle(n::Integer) = randcycle(GLOBAL_RNG, n)
-
-"""
-    randcycle!([rng=GLOBAL_RNG,] A::Array{<:Integer})
-
-Construct in `A` a random cyclic permutation of length `length(A)`.
-The optional `rng` argument specifies a random number generator, see
-[Random Numbers](@ref).
-
-# Examples
-```jldoctest
-julia> randcycle!(MersenneTwister(1234), Vector{Int}(6))
-6-element Array{Int64,1}:
- 3
- 5
- 4
- 6
- 1
- 2
-```
-"""
-function randcycle!(r::AbstractRNG, a::Array{<:Integer})
-    n = length(a)
-    n == 0 && return a
-    @assert n <= Int64(2)^52
-    a[1] = 1
-    mask = 3
-    @inbounds for i = 2:n
-        j = 1 + rand_lt(r, i-1, mask)
-        a[i] = a[j]
-        a[j] = i
-        i == 1+mask && (mask = 2mask + 1)
-    end
-    return a
-end
-
-randcycle!(a::Array{<:Integer}) = randcycle!(GLOBAL_RNG, a)
-
-end # module
